imageEvaller/requirements.embed.txt
deepdanbooru==1.0.2
tensorflow-io==0.34.0
mysql-connector-python==8.2.0